fix: move feedback endpoint out of predictions
