+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "vlmrun",
-  "version": "0.2.22",
+  "version": "0.2.23",
   "description": "The official TypeScript library for the VlmRun API",
   "author": "VlmRun <support@vlmrun.com>",
   "main": "dist/index.js",
